+++ src/sampler.h
@@ -8,19 +8,19 @@ class SamplePacket;
 class Sampler
 {
 public:
-    static const int sSamplesPerPixel;
-    static const int sSamplesPerAxis;
+    static const unsigned short sSamplesPerPixel;
+    static const unsigned short sSamplesPerAxis;
     static const float sSubpixelStep;
     
-    explicit Sampler(const int width, const int height);
+    explicit Sampler(const unsigned short width, const unsigned short height);
 
     bool buildSamplePacket(SamplePacket* packet);
 
 private:
     explicit Sampler() { } // Don't use default constructor
 
-    int mWidth, mHeight;
-    int mCurrentX, mCurrentY;
+    unsigned short mWidth, mHeight;
+    unsigned short mCurrentX, mCurrentY;
     
     pthread_mutex_t mLock;
 };
